Fixing sonarqube bugs (merged if statements + some small cleanups)
Former-commit-id: a593cd2173aa200c86714f597d616de4a3da8fd0
Former-commit-id: 484113a4a05ecbb42c051aab974c58c6b3f6c4a1

+++ src/DeltaShell.Plugins.NetworkEditor.Gui/Forms/ChartEditors/ShapeDeletePointTool.cs
@@ -23,14 +23,11 @@ namespace DeltaShell.Plugins.NetworkEditor.Gui.Forms.ChartEditors
                         if (null != ShapeModifyTool.ShapeFeatureEditor)
                         {
                             IPoint tracker = ShapeModifyTool.ShapeSelectTool.GetTrackerAt(ShapeModifyTool.ShapeFeatureEditor, tmP);
-                            if (null != tracker)
+                            if (null != tracker && ShapeModifyTool.ShapeFeatureEditor.CanDeleteTracker(tracker))
                             {
-                                if (ShapeModifyTool.ShapeFeatureEditor.CanDeleteTracker(tracker))
-                                {
-                                    ShapeModifyTool.ShapeFeatureEditor.DeleteTracker(tracker);
-                                    ((ShapeFeatureEditor)ShapeModifyTool.ShapeFeatureEditor).ShapeFeature.Invalidate();
-                                    return;
-                                }
+                                ShapeModifyTool.ShapeFeatureEditor.DeleteTracker(tracker);
+                                ((ShapeFeatureEditor)ShapeModifyTool.ShapeFeatureEditor).ShapeFeature.Invalidate();
+                                return;
                             }
                         }
                         ShapeModifyTool.ShapeSelectTool.MouseEvent(kind, e, c);
@@ -40,13 +37,10 @@ namespace DeltaShell.Plugins.NetworkEditor.Gui.Forms.ChartEditors
                     if (null != ShapeModifyTool.ShapeFeatureEditor)
                     {
                         IPoint tracker = ShapeModifyTool.ShapeSelectTool.GetTrackerAt(ShapeModifyTool.ShapeFeatureEditor, tmP);
-                        if (null != tracker)
+                        if (null != tracker && ShapeModifyTool.ShapeFeatureEditor.CanDeleteTracker(tracker))
                         {
-                            if (ShapeModifyTool.ShapeFeatureEditor.CanDeleteTracker(tracker))
-                            {
-                                c = ShapeModifyTool.ShapeFeatureEditor.GetCursor(tracker);
-                                return;
-                            }
+                            c = ShapeModifyTool.ShapeFeatureEditor.GetCursor(tracker);
+                            return;
                         }
                         c = Cursors.Default;
                     }
